+++ R/spacy_initialize.R
@@ -9,54 +9,68 @@
 #'   if \code{TRUE}, list available spaCy installations and prompt the user 
 #'   for which to use. If another (e.g. \code{python_executable}) is set, then 
 #'   this value will always be treated as \code{FALSE}.
+#' @param source_bash_profile logical; if \code{TRUE}, source \code{~/.bash_profile} before trying
+#'   to find python executables with spaCy installed. Most likely necessary to set \code{TRUE}, 
+#'   if using anaconda python in Mac or Linux. Default is \code{NULL} 
+#'   which functions as \code{TRUE} for non-Windows system (e.g. Mac/Linux) and 
+#'   \code{FALSE} for Windows system.
 #' @param virtualenv set a path to the python virtual environment with spaCy installed
 #'   Example: \code{virtualenv = "~/myenv"}
 #' @param condaenv set a path to the anaconda virtual environment with spaCy installed
 #'   Example: \code{condalenv = "myenv"}
+#' @param entity logical; if \code{FALSE} is selected, named entity recognition is turned off 
+#'   in spaCy. This will speed up the parsing as it will exclude \code{ner} from the pipeline. 
+#'   For details of spaCy pipeline, 
+#'   see \url{https://spacy.io/usage/processing-pipelines}. The option \code{FALSE} is available only 
+#'   for spaCy version 2.0.0 or higher.
 #' @export
 #' @author Akitaka Matsuo
 spacy_initialize <- function(model = "en", 
                              python_executable = NULL,
                              ask = FALSE,
+                             source_bash_profile = NULL,
                              virtualenv = NULL,
-                             condaenv = NULL) {
+                             condaenv = NULL, 
+                             entity = TRUE) {
 
     # here are a number of checkings
     if(!is.null(options("spacy_initialized")$spacy_initialized)){
         message("spaCy is already initialized")
         return(NULL)
     }
+    # set an option for source_bash_profile
+    if(is.null(source_bash_profile)) {
+        if(Sys.info()['sysname'] == "Windows"){
+            source_bash_profile <- FALSE
+        } else {
+            source_bash_profile <- TRUE
+        }
+    }
     # once python is initialized, you cannot change the python executables
     if(!is.null(options("python_initialized")$python_initialized)) {
-        message("Python space is already attached.  If you want to swtich to a different Python, please restart R.")
+        message("Python space is already attached.  If you want to switch to a different Python, please restart R.")
     } 
-    # a user can specify only one
-    else if(sum(!is.null(c(python_executable, virtualenv, condaenv))) > 1) {
-        stop(paste("Too many python environments are specified, please select only one",
-                   "from python_executable, virtualenv, and condaenv"))
+    else {
+        set_spacy_python_option(python_executable, 
+                                virtualenv, 
+                                condaenv, 
+                                ask, 
+                                source_bash_profile, 
+                                model)
     }
-    # give warning when nothing is specified
-    else if (sum(!is.null(c(python_executable, virtualenv, condaenv))) == 0){
-        # def_python <- ifelse(Sys.info()['sysname'] == "Windows", 
-        #                      system("where python", intern = TRUE), 
-        #                      system("which python", intern = TRUE))
-        message("Finding a python executable with spacy installed...")
-        spacy_python <- find_spacy(model, ask = ask)
-        if(!is.null(spacy_python)){
-            reticulate::use_python(spacy_python, required = TRUE)
-        } else {
-            stop("spaCy or language model ", model, " is not installed in any of python executables.")
-        }
-    } 
-    else {# set the path with reticulte
-        if(!is.null(python_executable)) {
-            if(check_spacy_model(python_executable, model) != "OK"){
-                stop("spaCy or language model ", model, " is not installed in ", python_executable)
+    
+    if (!is.null(options("spacy_python_setting")$spacy_python_setting)) {
+        ####
+        type <- options("spacy_python_setting")$spacy_python_setting$type
+        py_path <- options("spacy_python_setting")$spacy_python_setting$py_path
+        if(type == "python_executable") {
+            if(check_spacy_model(py_path, model) != "OK"){
+                stop("spaCy or language model ", model, " is not installed in ", py_path)
             }
-            reticulate::use_python(python_executable, required = TRUE)
+            reticulate::use_python(py_path, required = TRUE)
         }
-        else if(!is.null(virtualenv)) reticulate::use_virtualenv(virtualenv, required = TRUE)
-        else if(!is.null(condaenv)) reticulate::use_condaenv(condaenv, required = TRUE)
+        else if(type == "virtualenv") reticulate::use_virtualenv(py_path, required = TRUE)
+        else if(type == "condaenv") reticulate::use_condaenv(py_path, required = TRUE)
     }
     options("python_initialized" = TRUE) # next line could cause non-recoverable error 
     spacyr_pyexec(pyfile = system.file("python", "spacyr_class.py",
@@ -65,13 +79,18 @@ spacy_initialize <- function(model = "en",
     #     stop('value of lang option should be either "en" or "de"')
     # }
     spacyr_pyassign("model", model)
+    spacyr_pyassign("spacy_entity", entity)
+    options("spacy_entity" = entity)
     spacyr_pyexec(pyfile = system.file("python", "initialize_spacyPython.py",
                                        package = 'spacyr'))
     # spacy_version <- system2("pip", "show spacy", stdout = TRUE, stderr = TRUE)
     # spacy_version <- grep("Version" ,spacy_version, value = TRUE)
     # 
     spacy_version <- spacyr_pyget("versions")$spacy
-
+    if(entity == FALSE & as.integer(substr(spacy_version, 1, 1)) < 2){
+        message("entity == FALSE is only available for spaCy version 2.0.0 or higher")
+        options("spacy_entity" = TRUE)
+    }
     message("successfully initialized (spaCy Version: ", spacy_version,', language model: ', model, ')')
     options("spacy_initialized" = TRUE)
 }
@@ -104,22 +123,32 @@ spacy_finalize <- function() {
 #'   if \code{TRUE}, list available spaCy installations and prompt the user 
 #'   for which to use. If another (e.g. \code{python_executable}) is set, then 
 #'   this value will always be treated as \code{FALSE}.
+#' @param source_bash_profile logical; if \code{TRUE}, source \code{~/.bash_profile} before trying
+#'   to find python executables with spaCy installed. Most likely necessary to set \code{TRUE}, 
+#'   if using anaconda python in Mac or Linux. 
 #' @keywords internal
 #' @importFrom data.table data.table
-find_spacy <- function(model = "en", ask){
+find_spacy <- function(model = "en", ask, source_bash_profile){
     spacy_found <- `:=` <- NA
     spacy_python <- NULL
     options(warn = -1)
+    if(source_bash_profile == TRUE & Sys.info()['sysname'] == "Windows"){
+        message("the option, source_bash_profile == TRUE, will be ignored for windows system")
+    }
     py_execs <- if(Sys.info()['sysname'] == "Windows") {
         system2("where", "python", stdout = TRUE)
+    } else if(source_bash_profile == TRUE) {
+        c(system2('source', '~/.bash_profile; which -a python', stdout = TRUE),
+          system2('source', '~/.bash_profile; which -a python3', stdout = TRUE))
     } else {
         c(system2('which', '-a python', stdout = TRUE),
           system2('which', '-a python3', stdout = TRUE))
     }
+    py_execs <- unique(py_execs)
     options(warn = 0)
     
     if (length(py_execs) == 0 | grepl("not find", py_execs[1])[1]){
-        stop("No python was found on system PATH")
+        return(NA)
     }
     df_python_check <- data.table::data.table(py_execs, spacy_found = 0)
     for (i in 1:nrow(df_python_check)) {
@@ -161,7 +190,11 @@ find_spacy <- function(model = "en", ask){
 check_spacy_model <- function(py_exec, model) {
     options(warn = -1)
     py_exist <- if(Sys.info()['sysname'] == "Windows") {
-        system2("where", py_exec, stdout = TRUE)
+        if(py_exec %in% system2("where", "python", stdout = TRUE)) {
+            py_exec
+        } else {
+            NULL
+        }
     } else {
         system2('which', py_exec, stdout = TRUE)
     }
@@ -176,4 +209,56 @@ check_spacy_model <- function(py_exec, model) {
     })
     options(warn = 0)
     return(paste(sys_message, collapse = " "))
+}
+
+
+set_spacy_python_option <- function(python_executable = NULL, 
+                                    virtualenv = NULL, 
+                                    condaenv = NULL, 
+                                    ask = NULL, 
+                                    source_bash_profile = NULL,
+                                    model = NULL) {
+    # a user can specify only one
+    if(sum(!is.null(c(python_executable, virtualenv, condaenv))) > 1) {
+        stop(paste("Too many python environments are specified, please select only one",
+                   "from python_executable, virtualenv, and condaenv"))
+    }
+    # give warning when nothing is specified
+    else if (sum(!is.null(c(python_executable, virtualenv, condaenv))) == 1){
+        if(!is.null(python_executable)) {
+            if(check_spacy_model(python_executable, model) != "OK"){
+                stop("spaCy or language model ", model, " is not installed in ", python_executable)
+            }
+            options(spacy_python_setting = list(type = "python_executable",
+                                                py_path = python_executable))
+        }
+        else if(!is.null(virtualenv)) {
+            options(spacy_python_setting = list(type = "virtualenv",
+                                                py_path = virtualenv))
+            
+        }
+        else if(!is.null(condaenv)) {
+            options(spacy_python_setting = list(type = "condaenv",
+                                                py_path = condaenv))
+        }
+    }
+    else if (!is.null(options("spacy_python_setting")$spacy_python_setting)) {
+        message("python path is already set\nspacyr will use: ", 
+                options("spacy_python_setting")$spacy_python_setting$type, " = ",
+                options("spacy_python_setting")$spacy_python_setting$py_path)
+    }
+    else {
+        message("Finding a python executable with spacy installed...")
+        spacy_python <- find_spacy(model, ask = ask, 
+                                   source_bash_profile = source_bash_profile)
+        if (is.null(spacy_python)) {
+            stop("spaCy or language model ", model, " is not installed in any of python executables.")
+        } else if(is.na(spacy_python)) {
+            stop("No python was found on system PATH")
+        } else {
+            options(spacy_python_setting = list(type = "python_executable",
+                                                py_path = spacy_python))
+        }
+    }
+    return(NULL)
 }